Refactor: separate payment action saga
