+++ src/modules/dashboard/sagas/actions/index.ts
@@ -1,0 +1,7 @@
+import { all, call } from 'redux-saga/effects';
+
+import paymentActionSaga from './payment';
+
+export default function* actionsSagas() {
+  yield all([call(paymentActionSaga)]);
+}
