+++ src/modules/dashboard/sagas/index.ts
@@ -1,6 +1,7 @@
 import { all, call } from 'redux-saga/effects';
 
 import actionSagas from './actions';
+import TEMP_actionsSagas from './actions/index';
 import colonySagas from './colony';
 import colonyCreateSaga from './colonyCreate';
 import colonyDeploymentSaga from './colonyFinishDeployment';
@@ -10,6 +11,7 @@ import tokenSagas from './token';
 export default function* setupDashboardSagas() {
   yield all([
     call(actionSagas),
+    call(TEMP_actionsSagas),
     call(colonySagas),
     call(colonyCreateSaga),
     call(colonyDeploymentSaga),
